+++ CHANGELOG.md
@@ -6,8 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.2.1] Q1 2023
- - reduced actix-web workers to 1
- - refactored default address handling
+ - metrics_provider:
+   - reduced actix-web workers to 1
+   - refactored default address handling
+   - refactored drop behaviour
 
 ## [0.2.0] Q1 2023
  - replaced warp by actix-web